soc: move SoCController from soc_core to soc
